requirements.txt
Django==5.0.1
djangorestframework==3.14.0
celery==5.3.6
django-celery-results==2.5.1
flower==2.0.1
redis==5.0.1
python-dotenv==1.0.1
